requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# MongoDB async driver
motor==3.3.2
pymongo==4.6.1

# LangGraph for agent orchestration
langgraph
langchain
langchain-core

# LLM Providers
openai==1.12.0          # OpenAI API client
anthropic==0.18.1       # Optional: Anthropic Claude

# Pydantic for data validation
pydantic==2.6.0
pydantic-settings==2.1.0

# Environment variables
python-dotenv==1.0.0

# Utilities
httpx==0.26.0
aiofiles==23.2.1

# Testing
pytest
pytest-asyncio
pytest-cov
httpx

# Development
black
ruff
